ETSWORK-9564 - Add and configure react-i18next to support internationalization.
`react-18next` was the library chosen by Workgrid Engineering to use in our client applications. This initial configuration assumes we'll be able to fetch translations over the network at some point in the future or if we deploy as as webapp. In its present form for mobile the translations will be bundled with the application. Even though we're still using the blank Ionic template I've provided an example usage of the translation hook to demonstrate how it works. This is only temporary and it will be removed once we begin actual development of features.