ETSWORK-9303 - Bump Ionic Appflow Action for bugfix on branch builds
In testing PR checks I discovered that branch builds with Ionic Appflow were not working. I opened a support request with Ionic and they resolved the issue in `v1.0.2`. This bumps to that version and makes a few tweaks to allow iOS to be built on a branch as well.