ETSWORK-9967 - Implement authentication using `@ionic-enterprise/auth` and the components from `@workgrid/ui`. (#8)
This change implements authentication for Web, iOS and Android using the `@ionic-enterprise/auth` package. In addition to the React components that were necessary for this change native project configuration needed to be done to add the necessary redirect URIs. Furthermore custom hooks using Capacitor plugins were built to support the use case. Unfortunately the `@capacitor-community/react-hooks` couldn't be used because it's based on Capacitor 2 capabilities and my [PR](https://github.com/capacitor-community/react-hooks/pull/21) to upgrade it to version 3 has not been merged.

As part of trying to rebuild a robust test suite I bumped all deps and more importantly TypeScript to 4.x and `react-scripts` to 4.x. Without these `msw` would have to be pinned to an older version and we'd have to a specific jsdom implementation due to limitations in `react-scripts` 3.x.

The general pattern that's being established here is that we have an `AuthenticationProvider` component that understands the current authentication state as well as whether we have a company code and global config. If the company code is missing, global config is missing, or the user is currently not authenticated we display the sign page to acquire the proper user information. It's also worth noting that due to browser constraints we *must* show a Sign In button for the user to click to allow the browser popup to happen. If the user is authenticated the `AuthenticationProvider` renders its `children` exposing a `useAuth` hook to allow downstream components to retrieve an access token, sign out, or get the API host to use for network requests.

Since this PR is only addressing authentication there are still pieces missing such as space selection, feature flags, themes, etc. This work be done in subsequent PRs.