ETSWORK-10017 - Implement Space selection (#9)
The general pattern that's being established here is that we have a `SpacesProvider` component that understands whether the user has selected a space. If a current space has not been selected we display the space selection screen if a user has access to more than one space otherwise we choose the single space and render the `children` exposing the current space among other things via a custom hook.

Since this PR is only addressing spaces there are still pieces missing such as feature flags, themes, etc. This work be done in subsequent PRs.