ETSWORK-10034 - Add support for injecting themes at organization and space level. (#10)
With our use of Ionic our theme will be driven by [CSS variables](https://developer.mozilla.org/en-US/docs/Web/CSS/Using_CSS_custom_properties). Furthermore our plan is to remotely host organization and space level stylesheets for customers. This allows us to solve theming within the Workgrid App as shown in this PR as well as being able to inject the stylesheets for use by Microapps.

The implementation uses a top level `ThemeProvider` that exposes a custom `useTheme` hook with methods to set the `apiHost` and `spaceId`. This hook is used in `AuthenticationProvider` so we can theme once the global config is fetched and in the `SpacesProvider` when a space is chosen. This allows us to inject the theme at the earliest possible moment (sign in screen).